added some unit tests + gh workflow
