+++ src/main/java/me/duong/mrp/controller/MediaController.java
@@ -7,6 +7,7 @@ import me.duong.mrp.presentation.Controller;
 import me.duong.mrp.presentation.Method;
 import me.duong.mrp.presentation.Request;
 import me.duong.mrp.presentation.Responders;
+import me.duong.mrp.utils.Injector;
 import me.duong.mrp.utils.parser.DtoReader;
 import me.duong.mrp.utils.parser.DtoWriter;
 import me.duong.mrp.utils.parser.Guards;
@@ -15,7 +16,7 @@ public class MediaController {
     @Controller(path = "/api/media")
     public static void getAllMedia(Request request) {
         var filter = MediaFilter.fromQuery(request.query());
-        var service = new MediaService();
+        var service = Injector.INSTANCE.resolve(MediaService.class);
         var result = service.getAllMedia(filter, request.userId());
         String response = DtoWriter.writeJson(result);
         Responders.sendResponse(request, 200, response);
@@ -25,7 +26,7 @@ public class MediaController {
     public static void getMediaById(Request request) {
         var id = Guards.verifyWildcardInt(request, "id");
         if (id == -1) return;
-        var service = new MediaService();
+        var service = Injector.INSTANCE.resolve(MediaService.class);
         var result = service.getMediaById(id, request.userId());
         if (result.isPresent()) {
             String response = DtoWriter.writeJson(result.get());
@@ -43,7 +44,7 @@ public class MediaController {
         }
         var media = dto.get();
         media.setUserId(request.userId());
-        var service = new MediaService();
+        var service = Injector.INSTANCE.resolve(MediaService.class);
         var result = service.createMedia(media);
         String response = DtoWriter.writeJson(result);
         Responders.sendResponse(request, 201, response);
@@ -60,7 +61,7 @@ public class MediaController {
         var media = dto.get();
         media.setId(id);
         media.setUserId(request.userId());
-        var service = new MediaService();
+        var service = Injector.INSTANCE.resolve(MediaService.class);
         var result = service.updateMedia(media);
         if (result.isPresent()) {
             String response = DtoWriter.writeJson(result.get());
@@ -74,7 +75,7 @@ public class MediaController {
     public static void deleteMedia(Request request) {
         var id = Guards.verifyWildcardInt(request, "id");
         if (id == -1) return;
-        var service = new MediaService();
+        var service = Injector.INSTANCE.resolve(MediaService.class);
         service.deleteMedia(id, request.userId());
         Responders.sendResponse(request, 204);
     }
@@ -83,7 +84,7 @@ public class MediaController {
     public static void markFavoriteMedia(Request request) {
         var mediaId = Guards.verifyWildcardInt(request, "id");
         if (mediaId == -1) return;
-        var service = new MediaService();
+        var service = Injector.INSTANCE.resolve(MediaService.class);
         var result = service.markMediaAsFavorite(request.userId(), mediaId);
         Responders.sendResponse(request, result ? 200 : 400);
     }
@@ -92,7 +93,7 @@ public class MediaController {
     public static void unmarkFavoriteMedia(Request request) {
         var mediaId = Guards.verifyWildcardInt(request, "id");
         if (mediaId == -1) return;
-        var service = new MediaService();
+        var service = Injector.INSTANCE.resolve(MediaService.class);
         var result = service.unmarkMediaAsFavorite(request.userId(), mediaId);
         Responders.sendResponse(request, result ? 204 : 400);
     }
