+++ src/main/java/me/duong/mrp/service/MediaService.java
@@ -8,69 +8,18 @@ import me.duong.mrp.repository.RatingRepository;
 import java.util.List;
 import java.util.Optional;
 
-public class MediaService extends BaseService {
-    public Media createMedia(Media media) {
-        return super.callDbSession(session -> {
-            MediaRepository repository = new MediaRepository(session);
-            return repository.insertMedia(media);
-        });
-    }
+public interface MediaService {
+    Media createMedia(Media media);
 
-    public Optional<Media> updateMedia(Media media) {
-        return super.callDbSession(session -> {
-            MediaRepository repository = new MediaRepository(session);
-            if (repository.findMediaById(media.getId()).isEmpty()) {
-                return Optional.empty();
-            }
-            return Optional.of(repository.updateMedia(media));
-        });
-    }
+    Optional<Media> updateMedia(Media media);
 
-    public Optional<Media> getMediaById(int id, int loggedId) {
-        return super.callDbSession(session -> {
-            MediaRepository repository = new MediaRepository(session);
-            RatingRepository ratingRepository = new RatingRepository(session);
-            return repository.findMediaById(id).map(media ->
-                    media.setRatings(ratingRepository.findAllFilteredRatingsByMediaId(media.getId(), loggedId)));
-        });
-    }
+    Optional<Media> getMediaById(int id, int loggedId);
 
-    public List<Media> getAllMedia(MediaFilter filter, int loggedId) {
-        return super.callDbSession(session -> {
-            MediaRepository repository = new MediaRepository(session);
-            RatingRepository ratingRepository = new RatingRepository(session);
-            return repository.findAllMedia(filter).stream().map(media ->
-                            media.setRatings(ratingRepository.findAllFilteredRatingsByMediaId(media.getId(), loggedId)))
-                    .toList();
-        });
-    }
+    List<Media> getAllMedia(MediaFilter filter, int loggedId);
 
-    public void deleteMedia(int id, int userId) {
-        super.callDbSessionWithoutReturn(session -> {
-            MediaRepository repository = new MediaRepository(session);
-            repository.deleteMedia(id, userId);
-        });
-    }
+    void deleteMedia(int id, int userId);
 
-    public boolean markMediaAsFavorite(int userId, int mediaId) {
-        return super.callDbSession(session -> {
-            MediaRepository repository = new MediaRepository(session);
-            if (repository.checkAlreadyMarked(userId, mediaId)) {
-                return false;
-            }
-            repository.markMediaAsFavorite(userId, mediaId);
-            return true;
-        });
-    }
+    boolean markMediaAsFavorite(int userId, int mediaId);
 
-    public boolean unmarkMediaAsFavorite(int userId, int mediaId) {
-        return super.callDbSession(session -> {
-            MediaRepository repository = new MediaRepository(session);
-            if (!repository.checkAlreadyMarked(userId, mediaId)) {
-                return false;
-            }
-            repository.unmarkMediaAsFavorite(userId, mediaId);
-            return true;
-        });
-    }
+    boolean unmarkMediaAsFavorite(int userId, int mediaId);
 }
