+++ src/main/resources/database.sql
@@ -1,8 +1,8 @@
-DROP TABLE IF EXISTS users;
-DROP TABLE IF EXISTS media;
-DROP TABLE IF EXISTS rating;
-DROP TABLE IF EXISTS favorites;
-DROP TABLE IF EXISTS likes;
+DROP TABLE IF EXISTS users CASCADE;
+DROP TABLE IF EXISTS media CASCADE;
+DROP TABLE IF EXISTS ratings CASCADE;
+DROP TABLE IF EXISTS favorites CASCADE;
+DROP TABLE IF EXISTS likes CASCADE;
 
 CREATE TABLE IF NOT EXISTS users (
   id SERIAL PRIMARY KEY,
